Move check unit tests to jobs folder
